requirements.txt
langchain
langchain_community
langchain-groq
langchain_chroma
langchain-huggingface
sqlite3
pandas
pyprojroot
python-dotenv
streamlit
PyPDF2